전체 게시글 조회 API 수정 및 게시글 검색 API 추가

+++ SpringBootTemplate/src/main/java/com/example/demo/src/post/model/GetPostsRes.java
@@ -1,5 +1,6 @@
 package com.example.demo.src.post.model;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,5 +18,5 @@ public class GetPostsRes {
     private String tradeRegion; // 게시자의 지역
     private String postingTime; // 게시 시간 (ex: 3일전)
     private String payStatus; // 번개 페이 가능여부
-    private int likeNum; // 관심수
+    private long likeNum; // 관심수
 }
